Rename string_literal.hpp to literals.hpp
The string_ prefix was redundant with the folder name it is in.
