+++ fly/types/string/lexer.hpp
@@ -2,7 +2,7 @@
 
 #include "fly/types/string/detail/string_classifier.hpp"
 #include "fly/types/string/detail/string_traits.hpp"
-#include "fly/types/string/string_literal.hpp"
+#include "fly/types/string/literals.hpp"
 
 #include <cstdint>
 #include <optional>
