+++ test/types/string/format_string.cpp
@@ -6,7 +6,7 @@
 #include "fly/types/string/detail/format_string.hpp"
 
 #include "fly/types/numeric/literals.hpp"
-#include "fly/types/string/string_literal.hpp"
+#include "fly/types/string/literals.hpp"
 
 #include "catch2/catch_template_test_macros.hpp"
 #include "catch2/catch_test_macros.hpp"
