+++ test/types/string/string_format.cpp
@@ -5,8 +5,8 @@
 
 #include "fly/fly.hpp"
 #include "fly/types/numeric/literals.hpp"
+#include "fly/types/string/literals.hpp"
 #include "fly/types/string/string.hpp"
-#include "fly/types/string/string_literal.hpp"
 
 #include "catch2/catch_template_test_macros.hpp"
 #include "catch2/catch_test_macros.hpp"
