+++ test/types/string/string_traits.cpp
@@ -1,7 +1,7 @@
 #include "fly/types/string/detail/string_traits.hpp"
 
 #include "fly/traits/traits.hpp"
-#include "fly/types/string/string_literal.hpp"
+#include "fly/types/string/literals.hpp"
 
 #include "catch2/catch_template_test_macros.hpp"
 #include "catch2/catch_test_macros.hpp"
