+++ test/types/string/string_unicode.cpp
@@ -1,5 +1,5 @@
+#include "fly/types/string/literals.hpp"
 #include "fly/types/string/string.hpp"
-#include "fly/types/string/string_literal.hpp"
 
 #include "catch2/catch_template_test_macros.hpp"
 #include "catch2/catch_test_macros.hpp"
